Finished refactoring move generation

+++ chessEngine/chessEngine/BitBoardUtility.cpp
@@ -1,6 +1,6 @@
 #include "BitBoardUtility.h"
 
-bool BitBoardUtility::containsSquare(unsigned long bitBoard, int square)
+bool BitBoardUtility::containsSquare(unsigned long long bitBoard, int square)
 {
     return ((bitBoard >> square) & 1) != 0;
 }
